NodeDef edit: moved files into webapp/components/survey/NodeDefDetails
