+++ webapp/components/survey/NodeDefDetails/NodeDefDetails.js
@@ -1,65 +1,36 @@
-import './NodeDef.scss'
+import './NodeDefDetails.scss'
 
 import React from 'react'
 import { useDispatch } from 'react-redux'
-import { useHistory } from 'react-router'
 
 import * as StringUtils from '@core/stringUtils'
-import * as Survey from '@core/survey/survey'
 import * as NodeDef from '@core/survey/nodeDef'
-import * as NodeDefLayout from '@core/survey/nodeDefLayout'
 import * as Validation from '@core/validation/validation'
 
 import { useI18n } from '@webapp/store/system'
-import { NodeDefsActions, useSurvey, useSurveyCycleKey } from '@webapp/store/survey'
-import { DialogConfirmActions } from '@webapp/store/ui'
+import { NodeDefsActions } from '@webapp/store/survey'
 
 import TabBar from '@webapp/components/tabBar'
 import { FormItem, Input } from '@webapp/components/form/input'
 import * as NodeDefUiProps from '@webapp/loggedin/surveyViews/surveyForm/nodeDefs/nodeDefUIProps'
 
-import { useNodeDef } from './store/useNodeDef'
+import { useNodeDef } from './hooks'
 import * as NodeDefState from './store/state'
+import ButtonBar from './ButtonBar'
 import ValidationsProps from './advanced/ValidationsProps'
 import AdvancedProps from './advanced/AdvancedProps'
 import BasicProps from './basic/BasicProps'
 
-const _isNodeDefKeyEditDisabled = (survey, nodeDef) =>
-  !nodeDef ||
-  NodeDef.isRoot(nodeDef) ||
-  NodeDef.isMultiple(nodeDef) ||
-  (!NodeDef.isKey(nodeDef) &&
-    Survey.getNodeDefKeys(Survey.getNodeDefParent(nodeDef)(survey))(survey).length >= NodeDef.maxKeyAttributes) ||
-  NodeDef.isReadOnly(nodeDef)
-
-const _isNodeDefMultipleEditDisabled = (survey, surveyCycleKey, nodeDef) =>
-  !nodeDef ||
-  NodeDef.isPublished(nodeDef) ||
-  NodeDef.isKey(nodeDef) ||
-  NodeDefLayout.isRenderTable(surveyCycleKey)(nodeDef) ||
-  Survey.isNodeDefParentCode(nodeDef)(survey) ||
-  NodeDef.isReadOnly(nodeDef) ||
-  NodeDef.isAnalysis(nodeDef)
-
-const NodeDefView = () => {
+const NodeDefDetails = () => {
   const i18n = useI18n()
-  const history = useHistory()
   const dispatch = useDispatch()
 
-  const survey = useSurvey()
-  const surveyCycleKey = useSurveyCycleKey()
-
-  const { nodeDefState, setNodeDefState, editingNodeDefFromDesigner } = useNodeDef()
+  const { nodeDefState, editingFromDesigner, nodeDefParent, keyEditDisabled, multipleEditDisabled } = useNodeDef()
 
   const nodeDef = NodeDefState.getNodeDef(nodeDefState)
   const validation = NodeDefState.getValidation(nodeDefState)
-  const isDirty = NodeDefState.isDirty(nodeDefState)
 
-  const nodeDefName = NodeDef.getName(nodeDef)
   const nodeDefType = NodeDef.getType(nodeDef)
-  const nodeDefParent = Survey.getNodeDefByUuid(NodeDef.getParentUuid(nodeDef))(survey)
-  const nodeDefKeyEditDisabled = _isNodeDefKeyEditDisabled(survey, nodeDef)
-  const nodeDefMultipleEditDisabled = _isNodeDefMultipleEditDisabled(survey, surveyCycleKey, nodeDef)
 
   return nodeDef ? (
     <>
@@ -87,9 +58,9 @@ const NodeDefView = () => {
                 props: {
                   nodeDef,
                   validation,
-                  nodeDefKeyEditDisabled,
-                  nodeDefMultipleEditDisabled,
-                  editingNodeDefFromDesigner,
+                  nodeDefKeyEditDisabled: keyEditDisabled,
+                  nodeDefMultipleEditDisabled: multipleEditDisabled,
+                  editingNodeDefFromDesigner: editingFromDesigner,
                   setNodeDefParentUuid: (...args) => dispatch(NodeDefsActions.setNodeDefParentUuid(...args)),
                   setNodeDefProp: (...args) => dispatch(NodeDefsActions.setNodeDefProp(...args)),
                   putNodeDefLayoutProp: (...args) => dispatch(NodeDefsActions.putNodeDefLayoutProp(...args)),
@@ -123,47 +94,11 @@ const NodeDefView = () => {
             ]}
           />
 
-          <div className="button-bar">
-            <button
-              type="button"
-              className="btn btn-cancel"
-              onClick={() =>
-                isDirty
-                  ? dispatch(
-                      DialogConfirmActions.showDialogConfirm({
-                        key: 'common.cancelConfirm',
-                        onOk: NodeDefsActions.cancelNodeDefEdits(history),
-                      })
-                    )
-                  : dispatch(NodeDefsActions.cancelNodeDefEdits(history))
-              }
-            >
-              {i18n.t(isDirty ? 'common.cancel' : 'common.back')}
-            </button>
-            <button
-              type="button"
-              className="btn btn-primary"
-              onClick={() => dispatch(NodeDefsActions.saveNodeDefEdits())}
-              aria-disabled={!isDirty || StringUtils.isBlank(nodeDefName)}
-            >
-              <span className="icon icon-floppy-disk icon-left icon-12px" />
-              {i18n.t('common.save')}
-            </button>
-            {!NodeDef.isRoot(nodeDef) && !NodeDef.isTemporary(nodeDef) && (
-              <button
-                type="button"
-                className="btn btn-danger btn-delete"
-                onClick={() => dispatch(NodeDefsActions.removeNodeDef(nodeDef, history))}
-              >
-                <span className="icon icon-bin2 icon-left icon-12px" />
-                {i18n.t('common.delete')}
-              </button>
-            )}
-          </div>
+          <ButtonBar nodeDefState={nodeDefState} />
         </div>
       </div>
     </>
   ) : null
 }
 
-export default NodeDefView
+export default NodeDefDetails
