+++ webapp/components/survey/NodeDefDetails/NodeDefDetails.scss
