+++ webapp/components/survey/NodeDefDetails/hooks/index.js
@@ -1,0 +1,1 @@
+export { useNodeDef } from './useNodeDef'
