--- webapp/views/App/views/NodeDef/index.js
@@ -1,1 +1,0 @@
-export { default } from './NodeDef'
